+++ examples/Assets/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,231 +1,272 @@
 {
   "images" : [
     {
-      "idiom" : "iphone",
       "size" : "20x20",
+      "idiom" : "iphone",
+      "filename" : "geometry.20x20@2x~iphone.png",
       "scale" : "2x"
     },
     {
-      "idiom" : "iphone",
       "size" : "20x20",
+      "idiom" : "iphone",
+      "filename" : "geometry.20x20@3x~iphone.png",
       "scale" : "3x"
     },
     {
-      "idiom" : "iphone",
       "size" : "29x29",
+      "idiom" : "iphone",
+      "filename" : "geometry.29x29@2x~iphone.png",
       "scale" : "2x"
     },
     {
-      "idiom" : "iphone",
       "size" : "29x29",
+      "idiom" : "iphone",
+      "filename" : "geometry.29x29@3x~iphone.png",
       "scale" : "3x"
     },
     {
-      "idiom" : "iphone",
       "size" : "40x40",
+      "idiom" : "iphone",
+      "filename" : "geometry.40x40@2x~iphone.png",
       "scale" : "2x"
     },
     {
-      "idiom" : "iphone",
       "size" : "40x40",
+      "idiom" : "iphone",
+      "filename" : "geometry.40x40@3x~iphone.png",
       "scale" : "3x"
     },
     {
-      "idiom" : "iphone",
       "size" : "60x60",
+      "idiom" : "iphone",
+      "filename" : "geometry.60x60@2x~iphone.png",
       "scale" : "2x"
     },
     {
-      "idiom" : "iphone",
       "size" : "60x60",
+      "idiom" : "iphone",
+      "filename" : "geometry.60x60@3x~iphone.png",
       "scale" : "3x"
     },
     {
-      "idiom" : "ipad",
       "size" : "20x20",
+      "idiom" : "ipad",
+      "filename" : "geometry.20x20@1x~ipad.png",
       "scale" : "1x"
     },
     {
-      "idiom" : "ipad",
       "size" : "20x20",
+      "idiom" : "ipad",
+      "filename" : "geometry.20x20@2x~ipad.png",
       "scale" : "2x"
     },
     {
-      "idiom" : "ipad",
       "size" : "29x29",
+      "idiom" : "ipad",
+      "filename" : "geometry.29x29@1x~ipad.png",
       "scale" : "1x"
     },
     {
-      "idiom" : "ipad",
       "size" : "29x29",
+      "idiom" : "ipad",
+      "filename" : "geometry.29x29@2x~ipad.png",
       "scale" : "2x"
     },
     {
-      "idiom" : "ipad",
       "size" : "40x40",
+      "idiom" : "ipad",
+      "filename" : "geometry.40x40@1x~ipad.png",
       "scale" : "1x"
     },
     {
-      "idiom" : "ipad",
       "size" : "40x40",
+      "idiom" : "ipad",
+      "filename" : "geometry.40x40@2x~ipad.png",
       "scale" : "2x"
     },
     {
-      "idiom" : "ipad",
       "size" : "76x76",
+      "idiom" : "ipad",
+      "filename" : "geometry.76x76@1x~ipad.png",
       "scale" : "1x"
     },
     {
-      "idiom" : "ipad",
       "size" : "76x76",
+      "idiom" : "ipad",
+      "filename" : "geometry.76x76@2x~ipad.png",
       "scale" : "2x"
     },
     {
       "size" : "83.5x83.5",
       "idiom" : "ipad",
+      "filename" : "geometry.83.5x83.5@2x~ipad.png",
       "scale" : "2x"
     },
     {
       "size" : "1024x1024",
       "idiom" : "ios-marketing",
+      "filename" : "geometry.1024x1024@1x~ios-marketing.png",
       "scale" : "1x"
     },
     {
       "size" : "60x60",
       "idiom" : "car",
+      "filename" : "geometry.60x60@2x~car.png",
       "scale" : "2x"
     },
     {
       "size" : "60x60",
       "idiom" : "car",
+      "filename" : "geometry.60x60@3x~car.png",
       "scale" : "3x"
     },
     {
       "size" : "24x24",
-      "idiom" : "watch",
-      "scale" : "2x",
       "role" : "notificationCenter",
-      "subtype" : "38mm"
+      "filename" : "geometry.24x24@2x~watch.png",
+      "idiom" : "watch",
+      "subtype" : "38mm",
+      "scale" : "2x"
     },
     {
       "size" : "27.5x27.5",
-      "idiom" : "watch",
-      "scale" : "2x",
       "role" : "notificationCenter",
-      "subtype" : "42mm"
+      "filename" : "geometry.27.5x27.5@2x~watch.png",
+      "idiom" : "watch",
+      "subtype" : "42mm",
+      "scale" : "2x"
     },
     {
       "size" : "29x29",
       "idiom" : "watch",
+      "filename" : "geometry.29x29@2x~watch.png",
       "role" : "companionSettings",
       "scale" : "2x"
     },
     {
       "size" : "29x29",
       "idiom" : "watch",
+      "filename" : "geometry.29x29@3x~watch.png",
       "role" : "companionSettings",
       "scale" : "3x"
     },
     {
       "size" : "40x40",
-      "idiom" : "watch",
-      "scale" : "2x",
       "role" : "appLauncher",
-      "subtype" : "38mm"
+      "filename" : "geometry.40x40@2x~watch.png",
+      "idiom" : "watch",
+      "subtype" : "38mm",
+      "scale" : "2x"
     },
     {
       "size" : "44x44",
-      "idiom" : "watch",
-      "scale" : "2x",
       "role" : "appLauncher",
-      "subtype" : "40mm"
+      "filename" : "geometry.44x44@2x~watch.png",
+      "idiom" : "watch",
+      "subtype" : "40mm",
+      "scale" : "2x"
     },
     {
       "size" : "50x50",
-      "idiom" : "watch",
-      "scale" : "2x",
       "role" : "appLauncher",
-      "subtype" : "44mm"
+      "filename" : "geometry.50x50@2x~watch.png",
+      "idiom" : "watch",
+      "subtype" : "44mm",
+      "scale" : "2x"
     },
     {
       "size" : "86x86",
-      "idiom" : "watch",
-      "scale" : "2x",
       "role" : "quickLook",
-      "subtype" : "38mm"
+      "filename" : "geometry.86x86@2x~watch.png",
+      "idiom" : "watch",
+      "subtype" : "38mm",
+      "scale" : "2x"
     },
     {
       "size" : "98x98",
-      "idiom" : "watch",
-      "scale" : "2x",
       "role" : "quickLook",
-      "subtype" : "42mm"
+      "filename" : "geometry.98x98@2x~watch.png",
+      "idiom" : "watch",
+      "subtype" : "42mm",
+      "scale" : "2x"
     },
     {
       "size" : "108x108",
-      "idiom" : "watch",
-      "scale" : "2x",
       "role" : "quickLook",
-      "subtype" : "44mm"
+      "filename" : "geometry.108x108@2x~watch.png",
+      "idiom" : "watch",
+      "subtype" : "44mm",
+      "scale" : "2x"
     },
     {
       "size" : "1024x1024",
       "idiom" : "watch-marketing",
+      "filename" : "geometry.1024x1024@1x~watch-marketing.png",
       "scale" : "1x"
     },
     {
       "size" : "16x16",
       "idiom" : "mac",
+      "filename" : "geometry.16x16@1x~mac.png",
       "scale" : "1x"
     },
     {
       "size" : "16x16",
       "idiom" : "mac",
+      "filename" : "geometry.16x16@2x~mac.png",
       "scale" : "2x"
     },
     {
       "size" : "32x32",
       "idiom" : "mac",
+      "filename" : "geometry.32x32@1x~mac.png",
       "scale" : "1x"
     },
     {
       "size" : "32x32",
       "idiom" : "mac",
+      "filename" : "geometry.32x32@2x~mac.png",
       "scale" : "2x"
     },
     {
       "size" : "128x128",
       "idiom" : "mac",
+      "filename" : "geometry.128x128@1x~mac.png",
       "scale" : "1x"
     },
     {
       "size" : "128x128",
       "idiom" : "mac",
+      "filename" : "geometry.128x128@2x~mac.png",
       "scale" : "2x"
     },
     {
       "size" : "256x256",
       "idiom" : "mac",
+      "filename" : "geometry.256x256@1x~mac.png",
       "scale" : "1x"
     },
     {
       "size" : "256x256",
       "idiom" : "mac",
+      "filename" : "geometry.256x256@2x~mac.png",
       "scale" : "2x"
     },
     {
       "size" : "512x512",
       "idiom" : "mac",
+      "filename" : "geometry.512x512@1x~mac.png",
       "scale" : "1x"
     },
     {
       "size" : "512x512",
       "idiom" : "mac",
+      "filename" : "geometry.512x512@2x~mac.png",
       "scale" : "2x"
     }
   ],
   "info" : {
     "version" : 1,
-    "author" : "xcode"
+    "author" : "com.brightdigit.Speculid"
   }
 }
