+++ examples/Assets/Assets.xcassets/ImageSet.imageset/Contents.json
@@ -1,23 +1,27 @@
 {
   "images" : [
     {
+      "idiom" : "universal",
+      "filename" : "geometry.pdf"
+    },
+    {
+      "scale" : "1x",
+      "filename" : "geometry.1x~universal.png",
       "idiom" : "universal"
     },
     {
-      "idiom" : "universal",
-      "scale" : "1x"
+      "scale" : "2x",
+      "filename" : "geometry.2x~universal.png",
+      "idiom" : "universal"
     },
     {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
+      "scale" : "3x",
+      "filename" : "geometry.3x~universal.png",
+      "idiom" : "universal"
     }
   ],
   "info" : {
     "version" : 1,
-    "author" : "xcode"
+    "author" : "com.brightdigit.Speculid"
   }
 }
