+++ frameworks/speculid/Models/AssetSpecificationFile.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public struct AssetSpecificationFile: AssetSpecificationFileProtocol {
+  public let url: URL
+  public let document: AssetSpecificationDocumentProtocol
+}
