+++ frameworks/speculid/Models/AssetSpecificationFileProtocol.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol AssetSpecificationFileProtocol {
+  var url: URL { get }
+  var document: AssetSpecificationDocumentProtocol { get }
+}
