+++ frameworks/speculid/Protocols/AssetSpecificationDocumentProtocol.swift
@@ -1,5 +1,6 @@
 import Foundation
 
 public protocol AssetSpecificationDocumentProtocol: Codable {
+  var info: AssetSpecificationMetadataProtocol { get }
   var images: [AssetSpecificationProtocol] { get }
 }
