+++ frameworks/speculid/Protocols/AssetSpecificationMetadataProtocol.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public protocol AssetSpecificationMetadataProtocol: Codable {
+  var author: String { get }
+  var version: Int { get }
+}
